feat: Compatible with XPU monitoring

+++ agent/app/model/monitor.go
@@ -39,8 +39,8 @@ type MonitorGPU struct {
 	Temperature   int     `json:"temperature"`
 	PowerDraw     float64 `json:"powerDraw"`
 	MaxPowerLimit float64 `json:"maxPowerLimit"`
-	MemUsed       int     `json:"memUsed"`
-	MemTotal      int     `json:"memTotal"`
+	MemUsed       float64 `json:"memUsed"`
+	MemTotal      float64 `json:"memTotal"`
 	FanSpeed      int     `json:"fanSpeed"`
 	Processes     string  `json:"processes"`
 }
